cmd, core/state/snapshot: polish
