+++ core/state/snapshot/generate_test.go
@@ -170,9 +170,8 @@ func checkSnapRoot(t *testing.T, snap *diskLayer, trieRoot common.Hash) {
 	if snapRoot != trieRoot {
 		t.Fatalf("snaproot: %#x != trieroot #%x", snapRoot, trieRoot)
 	}
-	scanner := NewDanglingRange(snap.diskdb, nil, nil, false)
-	if len(scanner.result) != 0 {
-		t.Fatalf("Detected dangling storages %d", len(scanner.result))
+	if err := CheckDanglingStorage(snap.diskdb); err != nil {
+		t.Fatalf("Detected dangling storages %v", err)
 	}
 }
